Move logic out of SnakeYaml
